pyproject.toml
[project]
name = "quick-start"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Adrian Lyjak", email = "adrianlyjak@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "llama-index-workflows[server]>=1.3.0",
    "python-cowsay>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.llamadeploy.ui]
directory = "./ui"

[tool.llamadeploy]
app = "quick_start.workflow:app"
env-files = [".env"]
